requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
sqlalchemy==2.0.36
